+++ apps/daimo-bot/api/scripts/registerNeynarSigner.ts
