+++ apps/daimo-bot/api/types.ts
